Fix some GCC warnings
Fix some GCC warnings

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -56,8 +56,8 @@
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
 	int t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
-			(mask)) << (shift)), addr); \
+	writel_relaxed((t & ~(((u32) mask) << (shift))) + (((v) & \
+	((u32) mask)) << (shift)), addr); \
 } while (0)
